реализован модуль utils
